+++ tests/style/bug_191_constructor_naming/Kitten.m
@@ -1,0 +1,25 @@
+% (c) Copyright 2021 Florian Schanda
+
+classdef Kitten
+
+    properties
+        x (1, 1) uint32
+    end
+
+    methods
+
+        function obj = Kitten(x)
+            obj.x = x;
+        end
+
+    end
+
+    methods
+
+        function obj = Kitten(x)
+            obj.x = x;
+        end
+
+    end
+
+end