+++ tests/style/bug_191_constructor_naming/Potato.m
@@ -1,0 +1,17 @@
+% (c) Copyright 2021 Florian Schanda
+
+classdef Potato
+
+    properties
+        x (1, 1) uint32
+    end
+
+    methods
+
+        function obj = Potato(x)
+            obj.x = x;
+        end
+
+    end
+
+end